[Refactor/#113] remove id field in post dto
