+++ src/main/java/com/example/waggle/WaggleApplication.java
@@ -1,6 +1,6 @@
 package com.example.waggle;
 
-import com.example.waggle.global.security.SecurityUtil;
+import com.example.waggle.global.util.SecurityUtil;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
@@ -13,14 +13,14 @@ import java.util.Optional;
 @SpringBootApplication
 public class WaggleApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(WaggleApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(WaggleApplication.class, args);
+    }
 
 
-	@Bean
-	public AuditorAware<String> auditorProvider() {
-		return () -> Optional.of(SecurityUtil.getCurrentUsername());
-	}
+    @Bean
+    public AuditorAware<String> auditorProvider() {
+        return () -> Optional.of(SecurityUtil.getCurrentUsername());
+    }
 
 }
