+++ src/main/java/com/example/waggle/global/security/oauth2/handler/OAuth2AuthenticationSuccessHandler.java
@@ -1,5 +1,8 @@
 package com.example.waggle.global.security.oauth2.handler;
 
+import com.example.waggle.global.security.JwtToken;
+import com.example.waggle.global.security.TokenService;
+import com.example.waggle.global.util.CookieUtil;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -19,22 +22,22 @@ import java.io.IOException;
 public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
     @Value("${app.oauth2.authorizedRedirectUri}")
     private String redirectUri;
-//    private final TokenService tokenService;
+    private final TokenService tokenService;
 
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request,
                                         HttpServletResponse response,
                                         Authentication authentication) throws ServletException, IOException {
-//        JwtToken jwtToken = tokenService.generateToken(authentication);
-//        if (response.isCommitted()) {
-//            return;
-//        }
-//        String refreshToken = jwtToken.getRefreshToken();
-//        CookieUtil.addCookie(response, "refresh_token", refreshToken, 60);
-//        CookieUtil.deleteCookie(request, response, "JSESSIONID");
-//
-//        getRedirectStrategy().sendRedirect(request, response, redirectUri);
+        JwtToken jwtToken = tokenService.generateToken(authentication);
+        if (response.isCommitted()) {
+            return;
+        }
+        String refreshToken = jwtToken.getRefreshToken();
+        CookieUtil.addCookie(response, "refresh_token", refreshToken, 60);
+        CookieUtil.deleteCookie(request, response, "JSESSIONID");
+
+        getRedirectStrategy().sendRedirect(request, response, redirectUri);
     }
 
 }
