+++ src/main/java/com/example/waggle/web/dto/comment/CommentRequest.java
@@ -11,7 +11,6 @@ public class CommentRequest {
     @AllArgsConstructor
     @Builder
     public static class Post {
-        private Long id;
         private String content;
         private String username;
         private Long boardId;
