+++ src/main/java/com/example/waggle/web/dto/pet/PetRequest.java
@@ -16,7 +16,6 @@ public class PetRequest {
     @Builder
     @ToString
     public static class Post {
-        private Long id;
         @NotNull
         private String name;
         private String breed;
